Windows cross-build: fix Win32 portability, add build-win.sh + mingw toolchain
- platform.h: don't pull <windows.h> into consumers (collides with raylib
  DrawTextEx/CloseWindow/Rectangle); use concrete DWORD typedef
- platform.c: include <windows.h> + <sys/stat.h> for _S_IREAD/_S_IWRITE/_S_IEXEC
- tinyfiledialogs.h: include <stddef.h> for wchar_t
- add build-win.sh cross-compiling pdeide.exe/pde2c.exe for MinGW x86-64
  into build/win64-deps and packaging build/dist/pdeide-win64.zip
- add cmake/toolchains/mingw-w64-x86_64.cmake

+++ platform.h
@@ -11,8 +11,12 @@
 #define PDEIDE_PLATFORM_H
 
 #ifdef _WIN32
-#include <windows.h>
-typedef DWORD plat_pid_t;
+/* Do NOT include <windows.h> here: doing so leaks the Win32 SDK (winuser.h/
+ * windef.h) into every TU that includes platform.h, which collides with
+ * raylib identifiers (DrawTextEx, CloseWindow, Rectangle). The concrete
+ * DWORD is all consumers need; platform.c includes the real <windows.h>.
+ * On Win64 DWORD is an unsigned 32-bit type == unsigned long. */
+typedef unsigned long plat_pid_t;
 #else
 typedef long plat_pid_t;
 #endif
